+++ controllers/front/downloadl.php
@@ -17,7 +17,15 @@ if (!defined('_PS_VERSION_')) {
 
 use com\extremeidea\bidorbuy\storeintegrator\core as bobsi;
 
+/**
+ * Class BidorbuyStoreIntegratorDownloadLModuleFrontController.
+ */
 class BidorbuyStoreIntegratorDownloadLModuleFrontController extends ModuleFrontController {
+    /**
+     * Init Content 
+     * 
+     * @return void
+     */
     public function initContent() {
         parent::initContent();